LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-multiple-activation-expected.txt



FAIL CSS Values Test: ::scroll-marker activation of multiple markers assert_equals: first target ::scroll-marker is not active before activation expected "rgb(255, 0, 0)" but got ""
